Refactor story to be lesson prop on Game
