+++ src/components/organisms/Dialog.tsx
@@ -3,7 +3,7 @@ import styled from 'styled-components'
 import Form from '../molecules/Form'
 import Message from '../atoms/Message'
 import DialogNavigation from '../molecules/DialogNavigation'
-import story, { IStory } from '../../utils/story'
+import story, { IStory } from '../../lesons/story'
 import { State } from '../../state/storyInputs'
 
 const Container = styled.div`
